+++ ca/django_ca/management/mixins.py
@@ -46,7 +46,7 @@ class ArgumentsMixin(_Base, metaclass=abc.ABCMeta):
         """Add the --algorithm option."""
 
         default = constants.HASH_ALGORITHM_NAMES[type(ca_settings.CA_DIGEST_ALGORITHM)]
-        dsa_default = constants.HASH_ALGORITHM_NAMES[type(ca_settings.CA_DSA_DIGEST_ALGORITHM)]
+        dsa_default = constants.HASH_ALGORITHM_NAMES[type(ca_settings.CA_DSA_SIGNATURE_HASH_ALGORITHM)]
         parser.add_argument(
             "--algorithm",
             action=actions.AlgorithmAction,